Move providers out of index.ts
